feat(editor): display message when successfully edited or deleted session
